+++ src/Kooco.Pikachu.Domain/Items/ItemWithDetailsModel.cs
@@ -11,7 +11,7 @@ public class ItemWithDetailsModel
     public string? ItemName { get; set; }
     public string? Image { get; set; }
     public int? Stock { get; set; }
-    public float UnitPrice { get; set; }
+    public int UnitPrice { get; set; }
     public IEnumerable<CartItemDetailsModel> Details { get; set; }
 }
 
@@ -20,5 +20,5 @@ public class CartItemDetailsModel
     public Guid Id { get; set; }
     public string? Name { get; set; }
     public int? Stock { get; set; }
-    public float UnitPrice { get; set; }
+    public int UnitPrice { get; set; }
 }
